Added coursework from CS340
For the final dashboard to run, please download the module.py file, Project2FinalCs340.ipynb file, and have access to mongoDB, Download the AAC dataset and use the animals collection. You can also see a video of the functionality, along with a detailed ReadMe file, CS 340 README.docx to see the end result of my work.